@@ -1,4 +1,4 @@
-import {issue, issueCommand} from './command'
+import {issue, issueCommand, toCommandValue} from './command'
 
 import * as os from 'os'
 import * as path from 'path'
@@ -33,11 +33,13 @@ export enum ExitCode {
 /**
  * Sets env variable for this action and future actions in the job
  * @param name the name of the variable to set
- * @param val the value of the variable
+ * @param val the value of the variable. Non-string values will be converted to a string via JSON.stringify
  */
-export function exportVariable(name: string, val: string): void {
-  process.env[name] = val
-  issueCommand('set-env', {name}, val)
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function exportVariable(name: string, val: any): void {
+  const convertedVal = toCommandValue(val)
+  process.env[name] = convertedVal
+  issueCommand('set-env', {name}, convertedVal)
 }
 
 /**
@@ -78,10 +80,20 @@ export function getInput(name: string, options?: InputOptions): string {
  * Sets the value of an output.
  *
  * @param     name     name of the output to set
- * @param     value    value to store
+ * @param     value    value to store. Non-string values will be converted to a string via JSON.stringify
  */
-export function setOutput(name: string, value: string): void {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function setOutput(name: string, value: any): void {
   issueCommand('set-output', {name}, value)
+}
+
+/**
+ * Enables or disables the echoing of commands into stdout for the rest of the step.
+ * Echoing is disabled by default if ACTIONS_STEP_DEBUG is not set.
+ *
+ */
+export function setCommandEcho(enabled: boolean): void {
+  issue('echo', enabled ? 'on' : 'off')
 }
 
 //-----------------------------------------------------------------------
@@ -93,14 +105,22 @@ export function setOutput(name: string, value: string): void {
  * When the action exits it will be with an exit code of 1
  * @param message add error issue message
  */
-export function setFailed(message: string): void {
+export function setFailed(message: string | Error): void {
   process.exitCode = ExitCode.Failure
+
   error(message)
 }
 
 //-----------------------------------------------------------------------
 // Logging Commands
 //-----------------------------------------------------------------------
+
+/**
+ * Gets whether Actions Step Debug is on or not
+ */
+export function isDebug(): boolean {
+  return process.env['RUNNER_DEBUG'] === '1'
+}
 
 /**
  * Writes debug message to user log
@@ -112,18 +132,18 @@ export function debug(message: string): void {
 
 /**
  * Adds an error issue
- * @param message error issue message
+ * @param message error issue message. Errors will be converted to string via toString()
  */
-export function error(message: string): void {
-  issue('error', message)
+export function error(message: string | Error): void {
+  issue('error', message instanceof Error ? message.toString() : message)
 }
 
 /**
  * Adds an warning issue
- * @param message warning issue message
+ * @param message warning issue message. Errors will be converted to string via toString()
  */
-export function warning(message: string): void {
-  issue('warning', message)
+export function warning(message: string | Error): void {
+  issue('warning', message instanceof Error ? message.toString() : message)
 }
 
 /**
@@ -182,9 +202,10 @@ export async function group<T>(name: string, fn: () => Promise<T>): Promise<T> {
  * Saves state for current action, the state can only be retrieved by this action's post job execution.
  *
  * @param     name     name of the state to store
- * @param     value    value to store
+ * @param     value    value to store. Non-string values will be converted to a string via JSON.stringify
  */
-export function saveState(name: string, value: string): void {
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function saveState(name: string, value: any): void {
   issueCommand('save-state', {name}, value)
 }
 
