@@ -52,10 +52,11 @@ describe('@actions/context', () => {
     })
   })
 
-  it('works with pullRequest payloads', () => {
+  it('works with pull_request payloads', () => {
     delete process.env.GITHUB_REPOSITORY
     context.payload = {
-      pullRequest: {number: 2},
+      // eslint-disable-next-line @typescript-eslint/camelcase
+      pull_request: {number: 2},
       repository: {owner: {login: 'user'}, name: 'test'}
     }
     expect(context.issue).toEqual({
