@@ -778,7 +778,7 @@ describe('mkdirP', () => {
     await io.mkdirP(getTestTemp())
     await fs.mkdir(rootPath)
     await fs.mkdir(realDirPath)
-    await fs.writeFile(realFilePath, 'test real_dir/file.txt contet')
+    await fs.writeFile(realFilePath, 'test real_dir/file.txt content')
     await createSymlinkDir(realDirPath, symlinkDirPath)
 
     await io.mkdirP(symlinkDirPath)
@@ -802,7 +802,7 @@ describe('mkdirP', () => {
     await io.mkdirP(getTestTemp())
     await fs.mkdir(rootPath)
     await fs.mkdir(realDirPath)
-    await fs.writeFile(realFilePath, 'test real_dir/file.txt contet')
+    await fs.writeFile(realFilePath, 'test real_dir/file.txt content')
     await createSymlinkDir(realDirPath, symlinkDirPath)
 
     const subDirPath = path.join(symlinkDirPath, 'sub_dir')
@@ -1171,9 +1171,9 @@ describe('which', () => {
       const originalPath = process.env['PATH']
       try {
         // modify PATH
-        process.env['PATH'] = `${process.env['PATH']}${
-          path.delimiter
-        }${testPath}`
+        process.env[
+          'PATH'
+        ] = `${process.env['PATH']}${path.delimiter}${testPath}`
 
         // find each file
         for (const fileName of Object.keys(files)) {
@@ -1244,9 +1244,9 @@ describe('which', () => {
       await fs.writeFile(notExpectedFilePath, '')
       const originalPath = process.env['PATH']
       try {
-        process.env['PATH'] = `${process.env['PATH']}${
-          path.delimiter
-        }${testPath}`
+        process.env[
+          'PATH'
+        ] = `${process.env['PATH']}${path.delimiter}${testPath}`
         expect(await io.which(fileName)).toBe(expectedFilePath)
       } finally {
         process.env['PATH'] = originalPath
